chore(ast): move declarators to seperate file
